backend/requirements.txt
imagehash==4.3.1
Pillow==10.3.0
opencv-python==4.9.0.80
python-dotenv==1.0.1
